Move command to separate file
